// src/app/Helper/models/Category.ts
export class Category {
    ID: number;
    Name: string;
    DepartmentName: string;
    Department: string;
    Description: string;
    CreatedAt?:Date;
    UpdatedAt?: Date;
    CreatedByUserID?: number;
    IsActive?: boolean;
    DepartmentID?: number;
    ClassificationID?: number;
    Classification: string;
    UpdatePermission: string;
    
}